binary_port: Refactor component tests
